combine the utils classes and move them to azure_base
