+++ components/cookbooks/azure/libraries/availability_set.rb
@@ -5,10 +5,10 @@ module AzureCompute
   class AvailabilitySet
     def initialize(compute_service)
       creds =
-        AzureCommon::AzureUtils.get_credentials(compute_service['tenant_id'],
-                                                compute_service['client_id'],
-                                                compute_service['client_secret']
-                                               )
+        Utils.get_credentials(compute_service['tenant_id'],
+                              compute_service['client_id'],
+                              compute_service['client_secret']
+                             )
 
       @client = Azure::ARM::Compute::ComputeManagementClient.new(creds)
       @client.subscription_id = compute_service['subscription']
