+++ components/cookbooks/azure/recipes/add_net_sec_group.rb
@@ -1,6 +1,6 @@
 require File.expand_path('../../libraries/network_security_group.rb', __FILE__)
 require File.expand_path('../../../azure_base/libraries/logger.rb', __FILE__)
-require File.expand_path('../../libraries/azure_utils.rb', __FILE__)
+require File.expand_path('../../../azure_base/libraries/utils.rb', __FILE__)
 require File.expand_path('../../libraries/resource_group.rb', __FILE__)
 
 ::Chef::Recipe.send(:include, AzureNetwork)
@@ -8,7 +8,7 @@ require File.expand_path('../../libraries/resource_group.rb', __FILE__)
 ::Chef::Recipe.send(:include, Azure::ARM::Network::Models)
 
 #set the proxy if it exists as a cloud var
-AzureCommon::AzureUtils.set_proxy(node[:workorder][:payLoad][:OO_CLOUD_VARS])
+Utils.set_proxy(node[:workorder][:payLoad][:OO_CLOUD_VARS])
 
 include_recipe 'azure::get_credentials'
 credentials = node['azureCredentials']
