+++ components/cookbooks/azure/recipes/azure_subscription_status.rb
@@ -1,6 +1,5 @@
 require 'azure_mgmt_resources'
 require 'json'
-require File.expand_path('../../libraries/azure_utils.rb', __FILE__)
 require File.expand_path('../../../azure_base/libraries/logger.rb', __FILE__)
 
 ::Chef::Recipe.send(:include, Azure::ARM::Resources)
