+++ components/cookbooks/azure/recipes/build_storage_profile_for_add_node.rb
@@ -1,6 +1,5 @@
 require 'azure_mgmt_compute'
 require 'azure_mgmt_storage'
-require File.expand_path('../../libraries/regions.rb', __FILE__)
 require File.expand_path('../../../azure_base/libraries/logger.rb', __FILE__)
 
 ::Chef::Recipe.send(:include, Azure::ARM::Compute)
@@ -179,7 +178,7 @@ Chef::Log.info("Location: #{location}")
 # alpha-numberic  no special characters between 9 and 24 characters
 
 # name needs to be globally unique, but it also needs to be per region.
-generated_name = "oostg" + node.workorder.box.ciId.to_s + AzureRegions::RegionName.abbreviate(location)
+generated_name = "oostg" + node.workorder.box.ciId.to_s + Utils.abbreviate_location(location)
 
 if generated_name.length > 22
   generated_name = generated_name.slice!(0..21)  #making sure we are not over the limit
