+++ components/cookbooks/azure/recipes/del_keypair.rb
@@ -1,7 +1,5 @@
-require File.expand_path('../../libraries/azure_utils.rb', __FILE__)
-
 #set the proxy if it exists as a cloud var
-AzureCommon::AzureUtils.set_proxy(node[:workorder][:payLoad][:OO_CLOUD_VARS])
+Utils.set_proxy(node[:workorder][:payLoad][:OO_CLOUD_VARS])
 
 # NOTE: There is no recipe for deleting Availability Set.
 # An Availability Set is created as part of a Resource Group.
