+++ components/cookbooks/azuredns/recipes/remove_old_aliases.rb
@@ -1,12 +1,12 @@
 require File.expand_path('../../libraries/record_set.rb', __FILE__)
-require File.expand_path('../../../azure/libraries/azure_utils.rb', __FILE__)
 require File.expand_path('../../libraries/dns.rb', __FILE__)
 require File.expand_path('../../../azure_base/libraries/logger.rb', __FILE__)
+require File.expand_path('../../../azure_base/libraries/utils.rb', __FILE__)
 
 ::Chef::Recipe.send(:include, AzureDns)
 
 #set the proxy if it exists as a cloud var
-AzureCommon::AzureUtils.set_proxy(node.workorder.payLoad.OO_CLOUD_VARS)
+Utils.set_proxy(node.workorder.payLoad.OO_CLOUD_VARS)
 
 # get platform resource group and availability set
 include_recipe 'azure::get_platform_rg_and_as'
