+++ components/cookbooks/azuretrafficmanager/recipes/delete.rb
@@ -1,8 +1,7 @@
 require File.expand_path('../../libraries/traffic_managers.rb', __FILE__)
-require File.expand_path('../../../azure/libraries/azure_utils.rb', __FILE__)
 
 #set the proxy if it exists as a cloud var
-AzureCommon::AzureUtils.set_proxy(node.workorder.payLoad.OO_CLOUD_VARS)
+Utils.set_proxy(node.workorder.payLoad.OO_CLOUD_VARS)
 
 nsPathParts = node['workorder']['rfcCi']['nsPath'].split('/')
 cloud_name = node['workorder']['cloud']['ciName']
